Disband `traits` module and move `SieveFactory` to `hazmat`
